requirements.txt
# Environment
python-dotenv>=1.0.0

# LiveKit voice agent
livekit>=0.14.0
livekit-agents>=0.8.0

# Speech
faster-whisper>=1.0.0
pyttsx3>=2.90
soundfile>=0.12.0

# Audio analysis (stress from voice)
librosa>=0.10.0
numpy>=1.24.0

# Text sentiment
textblob>=0.17.0

# LLM (response generation)
groq>=0.4.0
